% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Absolute_correlation_matrix.R
\name{Absolute_correlation_matrix}
\alias{Absolute_correlation_matrix}
\title{Generation of absolute correlation matrixes}
\usage{
Absolute_correlation_matrix(input_se = input_se, list_gs = list_gs)
}
\arguments{
\item{input_se}{A SummarizedExperiment object}

\item{list_gs}{A list of gene set collection}
}
\value{
A list of matrices
}
\description{
Generation of absolute correlation matrices
}
\examples{

## One Gene Set

library(airway)
data(airway)

Signature <- list(GS1=names(airway@rowRanges)[1:6])

Mtx_1 <- Absolute_correlation_matrix(airway,Signature)

## Two or more Gene Sets

data(airway)
Signatures <- list(GS1=names(airway@rowRanges)[1:6],GS2=names(airway@rowRanges)[7:12])

Mtx_2 <- Absolute_correlation_matrix(airway,Signatures)

}
