% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_to_se.R
\name{experiment_to_se}
\alias{experiment_to_se}
\title{Creation of a SummarizedExperiment object from a gene expression data set and its metadata table}
\usage{
experiment_to_se(gematrix, metadata)
}
\arguments{
\item{gematrix}{A matrix}

\item{metadata}{Metadata of gematrix in data frame format}
}
\value{
summarized experiment object
}
\description{
Creation of a SummarizedExperiment object from a gene expression data set and its metadata table
}
\details{
The gematrix is a gene expression data frame where gene Symbols are in the first column. Sample names of gematrix must be in the first row. 
The metadata is a data frame where sample names are in the first column.
}
\examples{

library(SummarizedExperiment)

gematrix <- as.data.frame(matrix(rnorm(120),nrow=20))
colnames(gematrix) <- c("Sample1","Sample2","Sample3","Sample4","Sample5","Sample6")
gematrix <- cbind(seq(1,20,1), gematrix)
metadata <- data.frame(Sample=c("Sample1","Sample2","Sample3","Sample4","Sample5","Sample6"),
Treatment=c("Ctrl","Ctrl","Ctrl","Treated","Treated","Treated"))

se <- experiment_to_se(gematrix, metadata)
}
