% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsadapt.R
\name{gsadapt}
\alias{gsadapt}
\title{Gene Set Adaptation}
\usage{
gsadapt(se, gs_df, mggcc, ncec, plot_graph = c(FALSE, TRUE))
}
\arguments{
\item{se}{A summarizedExperiment}

\item{gs_df}{Dataframe with information of gene_symbol and signature name}

\item{mggcc}{Median gene-gene correlation Cut-off Numeric value between 0 to 1}

\item{ncec}{Node Centrality Eigenvalue Cut-off. Numeric value between 0 to 1}

\item{plot_graph}{Whether to plot graphs in directory}
}
\value{
dataframe with adapted gene sets
}
\description{
Adaptation of gene set collections to transcriptomic data
}
\details{
gsadapt
}
\examples{

### Two or more signatures (permissive filter)
library(SummarizedExperiment)
library(airway)

data(airway, package="airway")

Signature <- data.frame(Gene_ID = head(rownames(assay(airway)), n=20), 
Gene_Set = c(rep("GS1",10),rep("GS2",10))) 

DataMeta <- gsadapt(se = airway, gs_df=Signature, mggcc=0.2, ncec=0.1, plot_graph=FALSE)

### Two or more signatures (restrictive filter)
library(SummarizedExperiment)
library(airway)

data(airway, package="airway")

Signature <- data.frame(Gene_ID = head(rownames(assay(airway)), n=20), 
Gene_Set = c(rep("GS1",10),rep("GS2",10))) 

DataMeta <- gsadapt(se = airway, gs_df=Signature, mggcc=0.6, ncec=0.3, plot_graph=FALSE)
}
