% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsadapt_of_one_gs.R
\name{gsadapt_of_one_gs}
\alias{gsadapt_of_one_gs}
\alias{gsadapt}
\title{Gene Set Adaptation of one gene set}
\usage{
gsadapt(abs_mtx = abs_mtx, sig_name = sig_name, CCIM = CCIM, NCEC = NCEC)
}
\arguments{
\item{abs_mtx}{Absolute correlation matrix}

\item{sig_name}{The name of the gene set}

\item{CCIM}{Correlation cut-off}

\item{NCEC}{Centrality cut-off}
}
\value{
Dataframe with adapted genes
}
\description{
Gene Set Adaptation of one gene set
}
\details{
This function performs a simply gene set adaptation using graph models. CCIM and NCEC correspond to adaptation filters able to remove weak associated genes and poor connected genes into resultant graph g1.
}
\examples{

library("igraph")

Mtx <- matrix(rnorm(25),ncol=5)

Abs_mtx <- abs(cor(Mtx))
dimnames(Abs_mtx) <- list(c("Gene1","Gene2","Gene3","Gene4","Gene5"),
c("Gene1","Gene2","Gene3","Gene4","Gene5"))

Adapt_df <- gsadapt(abs_mtx=Abs_mtx,sig_name="GS1",CCIM=0.1,NCEC=0.2)

}
