% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_to_se.R
\name{Matrix_to_se}
\alias{Matrix_to_se}
\alias{Dataframe_to_se}
\title{Conversion of dataframe into SummarizedExperiment object}
\usage{
Dataframe_to_se(inp_df,chromosome_field,ranges,strand,id,condition)
}
\arguments{
\item{inp_df}{A matrix}

\item{chromosome_field}{The name of the column related to chromosome}

\item{ranges}{IRanges from each feature}

\item{strand}{The name of the column related to strand}

\item{id}{gene_id or feature name}

\item{condition}{Metadata of inp_df in dataframe format}
}
\value{
summarized experiment object
}
\description{
Conversion of dataframe into SummarizedExperiment object
}
\details{
Conversion using a expression matrix containing information about gene ranges and experimental condition of each samples of expression matrix. For creating genomic ranges chromosome location, strand, start and end gene locations are required to construct gene ranges. ColData is compulsory and should give enough information about sample condition. Names of samples must be identical between matrix colnames and colData sample names.
}
\examples{

library(GenomicRanges)
library(SummarizedExperiment)
library(S4Vectors)

matrix <- matrix(rnorm(120),nrow=20)
colnames(matrix) <- c("Sample1","Sample2","Sample3","Sample4","Sample5","Sample6")
rownames(matrix) <- seq(1,20,1)

chromosome_field <- rep("chr1",20)
ranges <- IRanges(floor(runif(20,1e5, 1e6)),width=100)
strand <- sample(c("+", "-","*"), 20, TRUE)
id <- rownames(matrix)

condition <- data.frame(Sample=c("Sample1","Sample2","Sample3","Sample4","Sample5","Sample6"),
Treatment=c("Ctrl","Ctrl","Ctrl","Treated","Treated","Treated"))

se <- Dataframe_to_se(matrix, chromosome_field, ranges, strand, id, condition)
}
