% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature_input_to_GSAdapt.R
\name{Signature_input_to_GSAdapt}
\alias{Signature_input_to_GSAdapt}
\title{Conversion of gene set collections}
\usage{
Signature_input_to_GSAdapt(Signature)
}
\arguments{
\item{Signature}{Dataframe with genes and signature columns}
}
\value{
List with signature information
}
\description{
Convert a dataframe in a list using signature name as factor and genes as features
}
\details{
Signature_input_to_GSAdapt

Function Signature_input_to_GSAdapt generates a list taking a two-column dataframe. It is supposed that features are in first column and signature name in second one. Features can be expressed as Ensembl ID, gene symbol or other annotated gene expression.
}
\examples{

## One Gene Set
Signature <- data.frame(GeneID = c("A","B"),Sig = rep("GS1",2))
Signature <- Signature_input_to_GSAdapt(Signature)

## Two and more Gene Sets
Signature <- data.frame(Gene_ID = c("A","B","C","D"),Sig = c(rep("GS1",2),rep("GS2",2)))
Signature <- Signature_input_to_GSAdapt(Signature)
}
